requirements.txt
acres==0.3.0
aiofiles==24.1.0
anyascii==0.3.2
anyio==4.9.0
arxiv==2.1.3
attrs==25.3.0
beautifulsoup4==4.13.3
bertopic==0.17.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.1
contractions==0.1.73
cycler==0.12.1
etelemetry==0.3.1
feedparser==6.0.11
filelock==3.13.1
Flask==3.1.0
fonttools==4.56.0
frontend==0.0.3
fsspec==2024.6.1
h11==0.14.0
hdbscan==0.8.40
httplib2==0.22.0
huggingface-hub==0.29.3
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
llvmlite==0.44.0
looseversion==1.3.0
lxml==5.3.1
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
narwhals==1.31.0
networkx==3.3
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numba==0.61.0
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.1.0
plotly==6.0.1
prov==2.0.1
puremagic==1.28
pyahocorasick==2.1.0
pyarrow==19.0.1
pycparser==2.22
pydot==3.0.4
PyMuPDF==1.25.4
pynndescent==0.5.13
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.29.0
sentence-transformers==3.4.1
setuptools==77.0.3
sgmllib3k==1.0.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.46.1
sympy==1.13.1
textsearch==0.0.24
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
traits==7.0.2
transformers==4.50.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
